+++ forms/locale/fr/LC_MESSAGES/django.po
@@ -1,0 +1,68 @@
+# SOME DESCRIPTIVE TITLE.
+# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER
+# This file is distributed under the same license as the PACKAGE package.
+# FIRST AUTHOR <EMAIL@ADDRESS>, YEAR.
+#
+#, fuzzy
+msgid ""
+msgstr ""
+"Project-Id-Version: \n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2024-06-25 10:02+0200\n"
+"PO-Revision-Date: 2024-06-25 10:09+0200\n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"Language: fr\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: nplurals=2; plural=(n > 1);\n"
+"X-Generator: Poedit 3.4.2\n"
+
+#: forms/models.py:14
+msgid "Single line text"
+msgstr "Texte sur une ligne"
+
+#: forms/models.py:15
+msgid "Multi-line text"
+msgstr "Texte sur plusieurs lignes"
+
+#: forms/models.py:16
+msgid "Email"
+msgstr "Adresse e-mail"
+
+#: forms/models.py:17
+msgid "Number"
+msgstr "Nombre"
+
+#: forms/models.py:18
+msgid "URL"
+msgstr "URL"
+
+#: forms/models.py:19
+msgid "Checkbox"
+msgstr "Case à cocher"
+
+#: forms/models.py:20
+msgid "Checkboxes"
+msgstr "Cases à cocher"
+
+#: forms/models.py:21
+msgid "Drop down"
+msgstr "Liste déroulante"
+
+#: forms/models.py:22
+msgid "Radio buttons"
+msgstr "Boutons radio"
+
+#: forms/models.py:23
+msgid "Date"
+msgstr "Date"
+
+#: forms/models.py:24
+msgid "Date/time"
+msgstr "Date et heure"
+
+#: forms/models.py:25
+msgid "Hidden field"
+msgstr "Champ caché"
