+++ forms/templates/forms/form_page.html
@@ -12,11 +12,14 @@
           {% csrf_token %}
           {% dsfr_form %}
           <input class="fr-btn" type="submit">
-          <ul>
-            {% for field in form %}
+
+          <!--
+            <ul>
+              {% for field in form %}
               <li>{{ field.label }}, {{ field.type }}, {{ field.field.widget.attrs }} {{ field.field.widget }}</li>
-            {% endfor %}
-          </ul>
+              {% endfor %}
+            </ul>
+          -->
         </form>
       </div>
     </div>
