+++ locale/fr/LC_MESSAGES/django.po
@@ -8,8 +8,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: \n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-06-04 15:08+0200\n"
-"PO-Revision-Date: 2024-06-04 15:16+0200\n"
+"POT-Creation-Date: 2024-06-25 10:02+0200\n"
+"PO-Revision-Date: 2024-06-25 10:10+0200\n"
 "Last-Translator: \n"
 "Language-Team: \n"
 "Language: fr\n"
@@ -22,7 +22,7 @@ msgstr ""
 #: templates/blocks/follow.html:9
 msgctxt "Section title"
 msgid "Subscribe to our newsletter"
-msgstr " Abonnez-vous à notre lettre d’information "
+msgstr "Abonnez-vous à notre lettre d’information"
 
 #: templates/blocks/follow.html:13
 msgctxt "Button title"
@@ -35,9 +35,7 @@ msgstr "S’abonner"
 
 #: templates/blocks/follow.html:30
 msgid "Follow us <br /> on social media"
-msgstr ""
-"Suivez-nous\n"
-"sur les réseaux sociaux "
+msgstr "Suivez-nous<br />sur les réseaux sociaux"
 
 #: templates/blocks/footer.html:5
 msgid "Back to home page"
